+++ test/components/previews/imagusu/design_system/core_components_component_preview.rb
@@ -1,0 +1,19 @@
+# frozen_string_literal: true
+
+module Imagusu
+  module DesignSystem
+    class CoreComponentsComponentPreview < ViewComponent::Preview
+      def default
+        render_with_template
+      end
+
+      def tones
+        render_with_template
+      end
+
+      def form_controls
+        render_with_template
+      end
+    end
+  end
+end