+++ test/engine_test.rb
@@ -13,8 +13,15 @@ class EngineTest < ActiveSupport::TestCase
     Rails.application.eager_load!
 
     assert_kind_of Class, Imagusu::DesignSystem::ButtonComponent
+    assert_kind_of Class, Imagusu::DesignSystem::AlertComponent
+    assert_kind_of Class, Imagusu::DesignSystem::BadgeComponent
+    assert_kind_of Class, Imagusu::DesignSystem::CardComponent
     assert_kind_of Class, Imagusu::DesignSystem::CheckboxComponent
+    assert_kind_of Class, Imagusu::DesignSystem::CheckboxGroupComponent
+    assert_kind_of Class, Imagusu::DesignSystem::ErrorSummaryComponent
     assert_kind_of Class, Imagusu::DesignSystem::FieldComponent
+    assert_kind_of Class, Imagusu::DesignSystem::FileUploadComponent
+    assert_kind_of Class, Imagusu::DesignSystem::LinkComponent
     assert_kind_of Class, Imagusu::DesignSystem::RadioGroupComponent
     assert_kind_of Class, Imagusu::DesignSystem::SelectComponent
     assert_kind_of Class, Imagusu::DesignSystem::TextAreaComponent